requirements.txt
streamlit
cohere
pandas
pytesseract
pdfplumber
Pillow